+++ include/domus/core/graph/graphs_algorithms.hpp
@@ -17,7 +17,7 @@ std::optional<Cycle> find_a_directed_cycle_in_graph(const Graph& graph);
 
 std::vector<Cycle> compute_cycle_basis(const Graph& graph);
 
-std::optional<std::vector<int>> make_topological_ordering(const Graph& graph);
+std::optional<std::vector<size_t>> make_topological_ordering(const Graph& graph);
 
 std::vector<Graph> compute_connected_components(const Graph& graph);
 
@@ -48,10 +48,10 @@ class Bipartition {
     Bipartition(Bipartition&&) noexcept;
     Bipartition& operator=(Bipartition&&) noexcept;
 
-    void set_side(int node_id, bool side);
-    bool get_side(int node_id) const;
-    bool has_node(int node_id) const;
-    bool are_in_same_side(int node_id_1, int node_id_2) const;
+    void set_side(size_t node_id, bool side);
+    bool get_side(size_t node_id) const;
+    bool has_node(size_t node_id) const;
+    bool are_in_same_side(size_t node_id_1, size_t node_id_2) const;
 };
 
 std::optional<Bipartition> compute_bipartition(const Graph& graph);
