+++ include/domus/core/graph/segment.hpp
@@ -18,8 +18,8 @@ struct Segment {
     Graph& get_segment();
     const Graph& get_segment() const;
     const NodesContainer& get_attachments() const;
-    bool has_attachment(int attachment_id) const;
-    void add_attachment(int attachment_id);
+    bool has_attachment(size_t attachment_id) const;
+    void add_attachment(size_t attachment_id);
     std::string to_string() const;
     void print() const;
     Segment();
@@ -29,5 +29,5 @@ std::vector<Segment> compute_segments(const Graph& graph, const Cycle& cycle);
 
 bool is_segment_a_path(const Segment& segment);
 
-std::deque<int>
-compute_path_between_attachments(const Segment& segment, int attachment_1, int attachment_2);
+std::deque<size_t>
+compute_path_between_attachments(const Segment& segment, size_t attachment_1, size_t attachment_2);
