+++ include/domus/orthogonal/equivalence_classes.hpp
@@ -14,13 +14,13 @@ class EquivalenceClasses {
     Int_ToInt_HashMap m_elem_to_class;
     Int_ToIntContainer_HashMap m_class_to_elems;
     IntHashSet m_all_classes;
-    bool has_class(int class_id) const;
+    bool has_class(size_t class_id) const;
 
   public:
-    void set_class(int elem, int class_id);
-    bool has_elem_a_class(int elem) const;
-    int get_class_of_elem(int elem) const;
-    const IntHashSet& get_elems_of_class(int class_id) const;
+    void set_class(size_t elem, size_t class_id);
+    bool has_elem_a_class(size_t elem) const;
+    size_t get_class_of_elem(size_t elem) const;
+    const IntHashSet& get_elems_of_class(size_t class_id) const;
     std::string to_string() const;
     void print() const;
     const IntHashSet& get_all_classes() const;
@@ -32,8 +32,8 @@ build_equivalence_classes(const Shape& shape, const Graph& graph);
 std::tuple<
     Graph,
     Graph,
-    std::map<std::pair<int, int>, std::pair<int, int>>,
-    std::map<std::pair<int, int>, std::pair<int, int>>>
+    std::map<std::pair<size_t, size_t>, std::pair<size_t, size_t>>,
+    std::map<std::pair<size_t, size_t>, std::pair<size_t, size_t>>>
 equivalence_classes_to_ordering(
     const EquivalenceClasses& equivalence_classes_x,
     const EquivalenceClasses& equivalence_classes_y,
